metodo create e findAll ok

+++ src/main/java/com/corso/dropwizard/udemy/dropbookamrksprova/core/User.java
@@ -2,7 +2,6 @@ package com.corso.dropwizard.udemy.dropbookamrksprova.core;
 
 import java.util.ArrayList;
 import java.util.List;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -11,7 +10,6 @@ import javax.persistence.NamedQueries;
 import javax.persistence.NamedQuery;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
-
 import lombok.Data;
 
 @Entity
@@ -26,7 +24,7 @@ public class User {
 	
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String username;
     private String password;
